+++ source/core/plugin/Container.h
