+++ source/core/plugin/Setting.h
@@ -13,7 +13,7 @@
 
 #include <core/plugin/setting_metadata.h>
 #include <core/plugin/Pattern.h>
-#include <core/plugin/container.h>
+#include <core/plugin/Container.h>
 #include <core/util/time_extensions.h>
 
 namespace DAQuiri {
