+++ tests/core/plugin/ContainerTest.cpp
@@ -1,4 +1,4 @@
-#include <core/plugin/container.h>
+#include <core/plugin/Container.h>
 #include <gtest/gtest.h>
 
 struct SomeType
